+++ selene-lib/tests/lints/standard_library/constants.lua
@@ -3,3 +3,5 @@ collectgarbage "count"
 
 collectgarbage("doge")
 collectgarbage(1)
+
+collectgarbage("coun" .. "t")